Test sqlite3 thread-safeness of LastInsertId
Essentially, should LastInsertId() return the last id inserted into the
database at all, or should it return the last id inserted after the
query, or set of queries, that it runs?
The answer appears to be that it returns the last id inserted when it
runs its query, and not any queries that occur after that.